# openapi/openapi.yaml
openapi: 3.0.3
info:
  title: Task API
  version: 1.0.0

paths:
  /tasks:
    post:
      summary: Create a new task
      tags:
        - Tasks
      operationId: createTask
      requestBody:
        description: The task to create
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TaskCreateInput'
      responses:
        '201':
          description: Task created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Task'
        '400':
          description: Bad request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

    get:
      summary: Get all tasks
      tags:
        - Tasks
      operationId: getTaskList
      responses:
        '200':
          description: List of tasks
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Task'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

  /tasks/{id}:
    put:
      summary: Update a task by ID
      tags:
        - Tasks
      operationId: updateTask
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TaskUpdateInput'
      responses:
        '200':
          description: Task updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Task'
        '400':
          description: Bad request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

    delete:
      summary: Delete a task by ID
      tags:
        - Tasks
      operationId: deleteTask
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      responses:
        '204':
          description: Successfully deleted (no content)
        '404':
          description: Task not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

components:
  schemas:

    Task:
      type: object
      properties:
        id:
          type: integer
          format: int64
          minimum: 0
        title:
          type: string
        is_done:
          type: boolean
      required:
            - id
            - title
            - is_done

    TaskUpdateInput:
      type: object
      properties:
        title:
          type: string
        is_done:
          type: boolean
          example: true
      required:
        - title
        - is_done

    TaskCreateInput:
      type: object
      properties:
        title:
          type: string
        is_done:
          type: boolean
          example: true
      required:
        - title
        - is_done

    Error:
      type: object
      properties:
        error:
          type: string
          example: "Invalid request"